python-model/requirements.txt
# Core NLP and Machine Learning
scikit-learn
nltk
pandas

# File Reading
PyPDF2
python-docx

# Visualization
wordcloud
matplotlib